PP-5358 Move event models to correct package
