+++ src/main/java/uk/gov/pay/directdebit/events/model/GoCardlessEventId.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.directdebit.payments.model;
+package uk.gov.pay.directdebit.events.model;
 
 import uk.gov.pay.commons.model.WrappedStringValue;
