+++ src/main/java/uk/gov/pay/directdebit/events/model/GoCardlessResourceType.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.directdebit.payments.model;
+package uk.gov.pay.directdebit.events.model;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
