+++ src/main/java/uk/gov/pay/directdebit/events/services/GoCardlessEventService.java
@@ -3,7 +3,7 @@ package uk.gov.pay.directdebit.events.services;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.gov.pay.directdebit.events.dao.GoCardlessEventDao;
-import uk.gov.pay.directdebit.payments.model.GoCardlessEvent;
+import uk.gov.pay.directdebit.events.model.GoCardlessEvent;
 
 import javax.inject.Inject;
