--- src/main/java/uk/gov/pay/directdebit/payments/model/GoCardlessResourceId.java
@@ -1,6 +1,0 @@
-package uk.gov.pay.directdebit.payments.model;
-
-public interface GoCardlessResourceId {
-
-    String getResourceId();
-}
