+++ src/main/java/uk/gov/pay/directdebit/webhook/gocardless/resources/WebhookGoCardlessResource.java
@@ -3,7 +3,7 @@ package uk.gov.pay.directdebit.webhook.gocardless.resources;
 import com.codahale.metrics.annotation.Timed;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import uk.gov.pay.directdebit.payments.model.GoCardlessEvent;
+import uk.gov.pay.directdebit.events.model.GoCardlessEvent;
 import uk.gov.pay.directdebit.webhook.gocardless.api.GoCardlessWebhookParser;
 import uk.gov.pay.directdebit.webhook.gocardless.services.WebhookGoCardlessService;
 import uk.gov.pay.directdebit.webhook.gocardless.support.WebhookVerifier;
