+++ src/main/java/uk/gov/pay/directdebit/webhook/gocardless/services/handlers/GoCardlessActionHandler.java
@@ -1,6 +1,6 @@
 package uk.gov.pay.directdebit.webhook.gocardless.services.handlers;
 
-import uk.gov.pay.directdebit.payments.model.GoCardlessEvent;
+import uk.gov.pay.directdebit.events.model.GoCardlessEvent;
 
 public interface GoCardlessActionHandler {
     void handle(GoCardlessEvent event);
