+++ src/main/java/uk/gov/pay/directdebit/webhook/gocardless/services/handlers/GoCardlessHandler.java
@@ -3,7 +3,7 @@ package uk.gov.pay.directdebit.webhook.gocardless.services.handlers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.gov.pay.directdebit.payments.model.DirectDebitEvent;
-import uk.gov.pay.directdebit.payments.model.GoCardlessEvent;
+import uk.gov.pay.directdebit.events.model.GoCardlessEvent;
 import uk.gov.pay.directdebit.events.services.GoCardlessEventService;
 import uk.gov.pay.directdebit.payments.services.PaymentService;
